Create SAS session configuration file

Create SAS session configuration file

Usage

sascfg(
  name = "default",
  host,
  sas,
  ssh = system("which ssh", intern = TRUE),
  encoding = "latin1",
  options = list("-fullstimer"),
  ...,
  sascfg = "sascfg_personal.py"
)

Arguments

name

(character)
name of the configuration.

host

(character)
host name of remote server.

sas

(character)
SAS executable path on remote server.

ssh

(character)
executable path of ssh.

encoding

(character)
encoding of the SAS session.

options

(list)
additional list of arguments to pass to ssh command.

...

additional arguments.

sascfg

(character)
target file of configuration.

Details

host and sas are required to connect to remote SAS server. Other arguments can follow default. If transferring datasets is needed, then tunnelling is required. Use tunnel = , rtunnel = to specify tunnels and reverse tunnels. The values should be length 1 integer.